created pipe, renamed property to plural
