+++ AdminWebsite/AdminWebsite/ClientApp/src/app/shared/shared.module.ts
@@ -21,6 +21,7 @@ import { VenuesMenuComponent } from './menus/venues-menu/venues-menu.component';
 import { MinutesToHoursPipe } from './pipes/minutes-to-hours.pipe';
 import { TooltipDirective } from './directives/tooltip.directive';
 import { SelectComponent } from './select';
+import { RolesToDisplayPipe } from './pipes/roles-to-display.pipe';
 
 @NgModule({
     imports: [
@@ -45,6 +46,7 @@ import { SelectComponent } from './select';
         JusticeUsersMenuComponent,
         VenuesMenuComponent,
         MinutesToHoursPipe,
+        RolesToDisplayPipe,
         TooltipDirective,
         SelectComponent
     ],
@@ -65,7 +67,8 @@ import { SelectComponent } from './select';
         VenuesMenuComponent,
         MinutesToHoursPipe,
         TooltipDirective,
-        SelectComponent
+        SelectComponent,
+        RolesToDisplayPipe
     ]
 })
 export class SharedModule {
