+++ AdminWebsite/AdminWebsite/ClientApp/src/app/work-allocation/manage-team/manage-team.component.ts
@@ -184,26 +184,6 @@ export class ManageTeamComponent implements OnInit, OnDestroy {
         this.userFormMode$.next('add');
     }
 
-    getRoles(roles: JusticeUserRole[]) {
-        let rolesToDisplay = '';
-        if (roles.includes(JusticeUserRole.Vho)) {
-            rolesToDisplay += 'CSO' + this.addComma(roles.length);
-        }
-        if (roles.includes(JusticeUserRole.VhTeamLead)) {
-            rolesToDisplay += 'ADMIN' + this.addComma(roles.length);
-        }
-        if (roles.includes(JusticeUserRole.StaffMember)) {
-            rolesToDisplay += 'SM';
-        }
-        return rolesToDisplay;
-    }
-
-    private addComma(length: number) {
-        if (length > 1) {
-            return ', ';
-        }
-        return '';
-    }
 }
 
 interface SearchForExistingJusticeUserForm {
